requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyrogram==2.0.106
tgcrypto==1.2.5
supabase==2.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
apscheduler==3.10.4
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

